move scene debug functions to own file
