+++ src/scene/scene.js
@@ -16,7 +16,6 @@ import StyleParser from '../styles/style_parser';
 import SceneLoader from './scene_loader';
 import View from './view';
 import Light from '../lights/light';
-import Tile from '../tile/tile';
 import TileManager from '../tile/tile_manager';
 import DataSource from '../sources/data_source';
 import '../sources/sources';
@@ -25,6 +24,7 @@ import RenderStateManager from '../gl/render_state';
 import CanvasText from '../styles/text/canvas_text';
 import FontManager from '../styles/text/font_manager';
 import MediaCapture from '../utils/media_capture';
+import setupSceneDebug from './scene_debug';
 
 // Load scene definition: pass an object directly, or a URL as string to load remotely
 export default class Scene {
@@ -98,7 +98,7 @@ export default class Scene {
         this.updating = 0;
         this.generation = Scene.generation; // an id that is incremented each time the scene config is invalidated
         this.last_complete_generation = Scene.generation; // last generation id with a complete view
-        this.setupDebug();
+        setupSceneDebug(this);
 
         this.log_level = options.logLevel || 'warn';
         log.setLevel(this.log_level);
@@ -481,7 +481,7 @@ export default class Scene {
             // Update feature selection map if necessary
             if (this.render_count !== this.last_render_count) {
                 this.render_count_changed = true;
-                this._logFirstFrame();
+                this.logFirstFrame();
 
                 this.getFeatureSelectionMapSize().then(size => {
                     this.selection_feature_count = size;
@@ -863,7 +863,7 @@ export default class Scene {
 
             // Profiling
             if (profile) {
-                this._profile('Scene.rebuild');
+                this.debug.profile('Scene.rebuild');
             }
 
             // Increment generation to ensure style/tile building stay in sync
@@ -892,7 +892,7 @@ export default class Scene {
         }).then(() => {
             // Profiling
             if (profile) {
-                this._profileEnd('Scene.rebuild');
+                this.debug.profileEnd('Scene.rebuild');
             }
         });
     }
@@ -905,7 +905,7 @@ export default class Scene {
         if (this.building) {
             log('info', 'Scene: build geometry finished');
             if (this.building.resolve) {
-                this._logFirstBuild();
+                this.logFirstBuild();
                 this.building.resolve(true);
             }
 
@@ -1316,22 +1316,8 @@ export default class Scene {
         return this.media_capture.stopVideoCapture();
     }
 
-
-    // Stats/debug/profiling methods
-
-    // Profile helpers, issues a profile on main thread & all workers
-    _profile(name) {
-        console.profile(`main thread: ${name}`); // eslint-disable-line no-console
-        WorkerBroker.postMessage(this.workers, 'self.profile', name);
-    }
-
-    _profileEnd(name) {
-        console.profileEnd(`main thread: ${name}`); // eslint-disable-line no-console
-        WorkerBroker.postMessage(this.workers, 'self.profileEnd', name);
-    }
-
     // Log first frame rendered (with any geometry)
-    _logFirstFrame() {
+    logFirstFrame() {
         if (this.last_render_count === 0 && !this.times.first_frame) {
             this.times.first_frame = (+new Date()) - this.start_time;
             log('debug', `Scene: initial frame time: ${this.times.first_frame}`);
@@ -1339,120 +1325,11 @@ export default class Scene {
     }
 
     // Log completion of first scene build
-    _logFirstBuild() {
+    logFirstBuild() {
         if (this.times.first_build == null) {
             this.times.first_build = (+new Date()) - this.start_time;
             log('debug', `Scene: initial build time: ${this.times.first_build}`);
         }
-    }
-
-    // Debug config and functions
-    setupDebug () {
-        let scene = this;
-        this.debug = {
-            // Rebuild geometry a given # of times and print average, min, max timings
-            timeRebuild (num = 1, options = {}) {
-                let times = [];
-                let cycle = () => {
-                    let start = +new Date();
-                    scene.rebuild(options).then(() => {
-                        times.push(+new Date() - start);
-
-                        if (times.length < num) {
-                            cycle();
-                        }
-                        else {
-                            let avg = ~~(times.reduce((a, b) => a + b) / times.length);
-                            log('info', `Profiled rebuild ${num} times: ${avg} avg (${Math.min(...times)} min, ${Math.max(...times)} max)`);
-                        }
-                    });
-                };
-                cycle();
-            },
-
-            // Return geometry counts of visible tiles, grouped by style name
-            geometryCountByStyle () {
-                let counts = {};
-                scene.tile_manager.getRenderableTiles().forEach(tile => {
-                    for (let style in tile.meshes) {
-                        counts[style] = counts[style] || 0;
-                        tile.meshes[style].forEach(mesh => {
-                            counts[style] += mesh.geometry_count;
-                        });
-                    }
-                });
-                return counts;
-            },
-
-            // Return geometry counts of visible tiles, grouped by base style name
-            geometryCountByBaseStyle () {
-                let style_counts = scene.debug.geometryCountByStyle();
-                let counts = {};
-                for (let style in style_counts) {
-                    let base = scene.styles[style].baseStyle();
-                    counts[base] = counts[base] || 0;
-                    counts[base] += style_counts[style];
-                }
-                return counts;
-            },
-
-            // Return sum of all geometry counts for visible tiles
-            geometryCountTotal () {
-                const styles = scene.debug.geometryCountByStyle();
-                return Object.keys(styles).reduce((p, c) => styles[c] + p, 0);
-            },
-
-            // Return geometry GL buffer sizes for visible tiles, grouped by style name
-            geometrySizeByStyle () {
-                let sizes = {};
-                scene.tile_manager.getRenderableTiles().forEach(tile => {
-                    for (let style in tile.meshes) {
-                        sizes[style] = sizes[style] || 0;
-                        tile.meshes[style].forEach(mesh => {
-                            sizes[style] += mesh.buffer_size;
-                        });
-                    }
-                });
-                return sizes;
-            },
-
-            // Return geometry GL buffer sizes for visible tiles, grouped by base style name
-            geometrySizeByBaseStyle () {
-                let style_sizes = scene.debug.geometrySizeByStyle();
-                let sizes = {};
-                for (let style in style_sizes) {
-                    let base = scene.styles[style].baseStyle();
-                    sizes[base] = sizes[base] || 0;
-                    sizes[base] += style_sizes[style];
-                }
-                return sizes;
-            },
-
-            // Return sum of all geometry GL buffer sizes for visible tiles
-            geometrySizeTotal () {
-                const styles = scene.debug.geometrySizeByStyle();
-                return Object.keys(styles).reduce((p, c) => styles[c] + p, 0);
-            },
-
-            // Return sum of all texture memory usage
-            textureSizeTotal() {
-                return Object.values(Texture.textures).map(t => t.byteSize()).reduce((p, c) => p + c);
-            },
-
-            layerStats () {
-                if (debugSettings.layer_stats) {
-                    return Tile.debugSumLayerStats(scene.tile_manager.getRenderableTiles());
-                }
-                else {
-                    log('warn', 'Enable the \'layer_stats\' debug setting to collect layer stats');
-                    return {};
-                }
-            },
-
-            renderableTilesCount () {
-                return scene.tile_manager.getRenderableTiles().length;
-            }
-        };
     }
 
 }
